+++ backend/src/helpers/DataAccess/attachmentUtils.ts
@@ -1,6 +1,6 @@
 import * as AWS from 'aws-sdk'
 import * as AWSXRay from 'aws-xray-sdk'
-import { createLogger } from '../utils/logger'
+import { createLogger } from '../../utils/logger'
 
 const XAWS = AWSXRay.captureAWS(AWS)
