+++ backend/src/lambda/http/deleteTodo.ts
@@ -4,7 +4,7 @@ import { APIGatewayProxyEvent, APIGatewayProxyResult } from 'aws-lambda'
 import * as middy from 'middy'
 import { cors, httpErrorHandler } from 'middy/middlewares'
 
-import { deleteTodo } from '../../helpers/todos'
+import { deleteTodo } from '../../helpers/BussinessLogic/todos'
 import { getUserId } from '../utils'
 
 export const handler = middy(
