+++ backend/src/lambda/http/generateUploadUrl.ts
@@ -3,8 +3,8 @@ import 'source-map-support/register'
 import { APIGatewayProxyEvent, APIGatewayProxyResult } from 'aws-lambda'
 import * as middy from 'middy'
 import { cors, httpErrorHandler } from 'middy/middlewares'
-import { AttachmentUtils } from '../../helpers/attachmentUtils'
-import { createAttachmentPresignedUrl } from '../../helpers/todos'
+import { AttachmentUtils } from '../../helpers/DataAccess/attachmentUtils'
+import { createAttachmentPresignedUrl } from '../../helpers/BussinessLogic/todos'
 import { getUserId } from '../utils'
 import * as uuid from 'uuid'
